services/fraud-detection-service/requirements.txt
flask==3.0.3
setuptools>=70.0.0
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation-flask==0.51b0
opentelemetry-semantic-conventions==0.51b0
